feat: Add CSV file upload endpoint and product listing endpoint
This commit adds a new endpoint to upload CSV files and insert their data into the database. It also includes a new endpoint to retrieve a list of products with pagination support. These endpoints enhance the functionality of the product router and provide the ability to manage product data in the API.

Note: This commit message is based on the analysis of the code changes and recent commits.